+++ kamehouse-admin/src/test/java/com/nicobrest/kamehouse/admin/model/kamehousecommand/WolKameHouseCommandTest.java
@@ -10,16 +10,16 @@ class WolKameHouseCommandTest extends AbstractKameHouseCommandTest {
 
   @Override
   protected KameHouseCommand getKameHouseCommand() {
-    return new WolKameHouseCommand("AA:BB:CC:DD:EE:FF", "192.168.0.255");
+    return new WolKameHouseCommand("AA:BB:CC:DD:EE:FF", "192.168.99.255");
   }
 
   @Override
   protected String getWindowsShellCommand() {
-    return "kamehouse/cmd/kamehouse-cmd.sh -o wol -mac AA:BB:CC:DD:EE:FF -broadcast 192.168.0.255";
+    return "kamehouse/cmd/kamehouse-cmd.sh -o wol -mac AA:BB:CC:DD:EE:FF -broadcast 192.168.99.255";
   }
 
   @Override
   protected String getLinuxShellCommand() {
-    return "kamehouse/cmd/kamehouse-cmd.sh -o wol -mac AA:BB:CC:DD:EE:FF -broadcast 192.168.0.255";
+    return "kamehouse/cmd/kamehouse-cmd.sh -o wol -mac AA:BB:CC:DD:EE:FF -broadcast 192.168.99.255";
   }
 }