+++ changelog.md
@@ -8,6 +8,7 @@
 #### v10.10 (vMaradona.Messi)
 
 - Added shell script and ui in groot to get the lists of successful and invalid desktop backgrounds processed
+- Updated default lan ips to 192.168.99.x
 - Added kamehouse-desktop tab in groot server manager
 - Added functionality to list successful and invalid backgrounds into list files in desktop
 - Updated default styles and position of desktop widgets